Time: 1 ms (100.00%) | Memory: 46.9 MB (15.79%) - LeetSync

+++ 3605-construct-the-minimum-bitwise-array-i/construct-the-minimum-bitwise-array-i.java
@@ -15,7 +15,7 @@ class Solution {
                     bit++;
                     temp = temp >> 1; 
                 }
-                ans[i]=num ^ (1 << bit-1);
+                ans[i]=(int)num ^ (1 << bit-1);
             }
         }
         return ans;
